Final update for the day. I need to troubleshoot db connectivity for each service
