--- microservices-ecommerce-platform/inventory-service/src/main/java/com/loyalblue/ecommerce/InventoryServiceApplication.java
@@ -1,13 +1,0 @@
-package com.loyalblue.ecommerce;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class InventoryServiceApplication
-{
-    public static void main(String[] args)
-    {
-        SpringApplication.run(InventoryServiceApplication.class, args);
-    }
-}
