--- microservices-ecommerce-platform/inventory-service/src/main/java/com/loyalblue/ecommerce/controller/HealthCheckController.java
@@ -1,24 +1,0 @@
-package com.loyalblue.ecommerce.controller;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class HealthCheckController {
-
-    @Autowired
-    private JdbcTemplate jdbcTemplate;
-
-    @GetMapping("/health")
-    public String healthCheck() {
-        try {
-            jdbcTemplate.execute("SELECT 1");
-            return "Database connection successful!";
-        } catch (Exception e) {
-            return "Database connection failed: " + e.getMessage();
-        }
-    }
-}
-
